add .roll command and tests
